backend/src/main/java/com/siddhu/spring/customer/CustomerRowMapper.java
package com.siddhu.spring.customer;

import org.hibernate.annotations.Comment;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;

import java.sql.ResultSet;
import java.sql.SQLException;

@Component
public class CustomerRowMapper implements RowMapper {
    @Override
    public Object mapRow(ResultSet rs, int rowNum) throws SQLException {
        return new Customer(
                rs.getInt("id"),
                rs.getString("name"),
                rs.getDate("date"),
                rs.getString("gender")
        );
    }
}
